+++ volmdlr/step.py
@@ -117,6 +117,7 @@ def item_defined_transformation(arguments, object_dict):
     # TODO : how to frame map properly from these two Frame3D ?
     # return volmdlr_object2 - volmdlr_object1
     return volmdlr_object2
+#
 
 
 def manifold_surface_shape_representation(arguments, object_dict):
@@ -188,13 +189,17 @@ def representation_relationship_representation_relationship_with_transformation_
         arguments, object_dict):
     if arguments[2] in object_dict:
         if isinstance(object_dict[arguments[2]], list):
+
             for shell3d in object_dict[arguments[2]]:
+
                 frame3d = object_dict[arguments[4]]
                 shell3d.frame_mapping_inplace(frame3d, 'old')
                 # return shell3d
             return None
+
         else:
             shell3d = object_dict[arguments[2]]
+
             frame3d = object_dict[arguments[4]]
             shell3d.frame_mapping_inplace(frame3d, 'old')
             # return shell3d
@@ -565,8 +570,6 @@ class Step(dc.DessiaObject):
 
     def to_volume_model(self, show_times=False):
         """
-        no_bug_mode=True loops on instanciate method's KeyErrors until all
-        the KeyErrors can be instanciated.
         show_times=True displays the numer of times a given class has been
         instanciated and the totatl time of all the instanciations of this
         given class.
@@ -590,7 +593,6 @@ class Step(dc.DessiaObject):
             if node != '#0' and self.functions[node].name == 'BREP_WITH_VOIDS':
                 shell_nodes.append(node)
                 not_shell_nodes.append(int(self.functions[node].arg[1][1:]))
-
         frame_mapped_shell_node = []
         for s_node in shell_nodes:
             for fm_node in frame_mapping_nodes:
@@ -604,14 +606,13 @@ class Step(dc.DessiaObject):
         for node in shell_nodes + frame_mapping_nodes:
             self.graph.add_edge('#0', node)
 
-        # self.draw_graph(self.graph, reduced=True, save=True)
+        # self.draw_graph(self.graph, reduced=True)
 
         nodes = []
         i = 1
         new_nodes = True
         while new_nodes:
-            new_nodes = list(nx.descendants_at_distance(
-                self.graph, '#0', i))[::-1]
+            new_nodes = list(nx.descendants_at_distance(self.graph, '#0', i))[::-1]
             nodes.extend(new_nodes)
             i += 1
 
@@ -627,9 +628,7 @@ class Step(dc.DessiaObject):
                     for instanciate_id in instanciate_ids[::-1]:
                         t = time.time()
                         volmdlr_object = self.instanciate(
-                            self.functions[instanciate_id].name,
-                            self.functions[instanciate_id].arg[:],
-                            object_dict)
+                            self.functions[instanciate_id].name, self.functions[instanciate_id].arg[:], object_dict)
                         t = time.time() - t
                         object_dict[instanciate_id] = volmdlr_object
                         if show_times:
@@ -656,8 +655,10 @@ class Step(dc.DessiaObject):
                 shells.extend(object_dict[node])
             else:
                 shells.append(object_dict[node])
-
-        return volmdlr.core.VolumeModel(shells)
+        volume_model = volmdlr.core.VolumeModel(shells)
+        # bounding_box = volume_model.bounding_box
+        # volume_model = volume_model.translation(-bounding_box.center)
+        return volume_model
 
     def to_points(self):
         object_dict = {}