+++ CHANGELOG.md
@@ -14,8 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Circle2D: split_by_line
 * BoundingRectangle: bounds, plot, area, center, b_rectangle_intersection, is_inside_b_rectangle, point_belongs, intersection_area, distance_to_b_rectangle, distance_to_point
 * Cylinder: random_point_inside, interference_volume_with_other_cylinder, lhs_points_inside
-* CylindricalSurface3D: line_intersections, linesegment_intersections
+* CylindricalSurface3D: line_intersections, linesegment_intersections, plane_intersection
 * Line2D: point_distance
+* Line3D: to_2d
+
+
 
 ### Fixed
 
@@ -27,12 +30,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * STL: from stream (use BinaryFile and StringFile instead of io.BinaryIO and FileIO)
 * step: from stream (use BinaryFile instead of io.BinaryIO)
 
+
 ### Performance improvements
 
 * Avoid unneeded bbox computation
 
 
 ### Refactorings
+
 
 
 
@@ -47,6 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * CylindricalSurface3D: line_intersections
 
 
+
 ## v0.6.0 [11/7/2022]
 
 ### New Features
@@ -59,7 +65,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * ElementsGroup: nodes, elements_per_node
 * Mesh: bounding_rectangle, delete_duplicated_nodes
 * PlaneFace3D: cut_by_coincident_face
-
+* Vector2D: to_step
+* BSplineCurve2D: to_step
+* LineSegment3D: to_bspline_curve
+* BSplineCurve3D: from_geomdl_curve
+* Surface2D: line_crossings
+* Surface2D: from_contour
+* BSplineSurface3D: simpifly_surface - verifies if BSplineSurface3D could be a Plane3D
+* OpenShell3D: to_step_face_ids
+* Contour2D: repair_cut_contour
+* Circle2D: cut_by_line
 
 ### Fixed
 
@@ -75,6 +90,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Contour2D: point_belongs (bug when contour has only one primitive, like FullArc2D)
 * Contour: contours_from_edges
 * PlaneFace3D: face_intersections
+* Edge: insert_knots_and_mutiplicity
+* BSplineCurve3D: from_step
+* Surface2D: cut_by_line
+* Circle3D: to_step
 
 
 ### Performance improvements